Unlink temporary files immediately after creation
Basic idea is that the temporary file will be automatically removed
whenever the file is closed, in case we execute some code path that
exits without cleaning up.  Likely not supported on MSDOS and VMS, but
these likely don't support mkstemp() or mkstemps() either and would need
a different solution entirely.

Replaced the other file deletions with remove() from the standard C
library, which doesn't necessarily offer this guarantee but hopefully
exists on all platforms.  We've closed the files first and don't need
that guarantee.